Roll back toml dependency to 0.1.28 so url-0.5.9 parses.
When cargo 0.10.0 was released the latest version of toml was 0.1.28,
which allows malformed TOML where table entries don't have a newline
between the header and the first key.

Toml 0.1.29 disabled this by default and later versions of cargo pass
a flag to get the compatibility mode.

Until the next version of cargo is released, roll back the toml dependency
for cargo-vendor to 0.1.28 to match the behavior of cargo 0.10.0.

Fixes: #6