Don't put versions in filenames by default
Most of the time it's not actually needed, and only in the rare case that we've
got multiple versions do we actually need a directory with a version in it. For
that case we can just put the older versions with version in filenames and newer
versions with just the crate name in the directory name.